+++ pkg/resources/database/resource_service_mysql.go
@@ -6,12 +6,13 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"strings"
 	"time"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/setvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
@@ -48,6 +49,9 @@ var ResourceMysqlSchema = schema.SingleNestedAttribute{
 			MarkdownDescription: "The automated backup schedule (`HH:MM`).",
 			Optional:            true,
 			Computed:            true,
+			PlanModifiers: []planmodifier.String{
+				stringplanmodifier.UseStateForUnknown(),
+			},
 		},
 		"ip_filter": schema.SetAttribute{
 			ElementType:         types.StringType,
@@ -62,13 +66,17 @@ var ResourceMysqlSchema = schema.SingleNestedAttribute{
 			MarkdownDescription: "MySQL configuration settings in JSON format (`exo dbaas type show mysql --settings=mysql` for reference).",
 			Optional:            true,
 			Computed:            true,
+			PlanModifiers: []planmodifier.String{
+				stringplanmodifier.UseStateForUnknown(),
+			},
 		},
 		"version": schema.StringAttribute{
 			MarkdownDescription: "MySQL major version (`exo dbaas type show mysql` for reference; may only be set at creation time).",
 			Optional:            true,
 			Computed:            true,
-			Validators: []validator.String{
-				validators.IsMajorVersionValidator,
+			PlanModifiers: []planmodifier.String{
+				versionUseStateUnlessChanged(),
+				stringplanmodifier.RequiresReplaceIfConfigured(),
 			},
 		},
 	},
@@ -258,7 +266,7 @@ pooling:
 	if data.Mysql.Version.IsUnknown() {
 		data.Mysql.Version = types.StringNull()
 		if apiService.Version != nil {
-			data.Mysql.Version = types.StringValue(strings.SplitN(*apiService.Version, ".", 2)[0])
+			data.Mysql.Version = types.StringValue(*apiService.Version)
 		}
 	}
 
@@ -343,7 +351,7 @@ func (r *ServiceResource) readMysql(ctx context.Context, data *ServiceResourceMo
 
 	data.Mysql.Version = types.StringNull()
 	if apiService.Version != nil {
-		data.Mysql.Version = types.StringValue(strings.SplitN(*apiService.Version, ".", 2)[0])
+		data.Mysql.Version = types.StringValue(*apiService.Version)
 	}
 
 	data.Mysql.Settings = types.StringNull()
@@ -532,7 +540,7 @@ func (r *ServiceResource) updateMysql(ctx context.Context, stateData *ServiceRes
 	if stateData.Mysql.Version.IsUnknown() {
 		stateData.Mysql.Version = types.StringNull()
 		if apiService.Version != nil {
-			stateData.Mysql.Version = types.StringValue(strings.SplitN(*apiService.Version, ".", 2)[0])
+			stateData.Mysql.Version = types.StringValue(*apiService.Version)
 		}
 	}