+++ CHANGELOG.md
@@ -10,6 +10,8 @@ IMPROVEMENTS:
 
 BUG FIXES:
 - fix: typo in security group rule import doc #559
+- dbaas: allows services with minor/patch version #563
+- improve acc test #563
 
 ## 0.69.3